MariaIvyGuese_COMP377Lab4/requirements.txt
langchain
langchain-community
langchain-google-genai
pymupdf
faiss-cpu
numpy
python-dotenv
